Rename leds to led
To match the Tock drivers rename the libtock-rs leds to led.
