+++ examples/blink.rs
@@ -9,7 +9,7 @@ libtock_core::stack_size! {0x400}
 async fn main() -> TockResult<()> {
     let mut drivers = libtock::retrieve_drivers()?;
 
-    let leds_driver = drivers.leds.init_driver()?;
+    let leds_driver = drivers.led.init_driver()?;
     let mut timer_driver = drivers.alarm.create_timer_driver();
     let timer_driver = timer_driver.activate()?;
