+++ src/lib.rs
@@ -16,7 +16,7 @@ pub mod gpio;
 pub mod hmac;
 pub mod i2c_master;
 pub mod i2c_master_slave;
-pub mod leds;
+pub mod led;
 pub mod result;
 pub mod rng;
 pub mod sensors;
